Move jpeg related code to external
