+++ rtengine/imageio.cc
@@ -33,6 +33,7 @@
 
 #include <fcntl.h>
 #include <glib/gstdio.h>
+#include <jpeg_ijg/jpeg.h>
 #include <png.h>
 #include <tiff.h>
 #include <tiffio.h>
@@ -47,7 +48,6 @@
 #include "iccjpeg.h"
 #include "imagedata.h"
 #include "imageio.h"
-#include "jpeg.h"
 #include "procparams.h"
 #include "rt_math.h"
 #include "settings.h"
